requirements.txt
aiomysql==0.2.0
aiosqlite==0.20.0
cryptography==39.0.1
fastapi
oauthlib==3.2.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0
pytest>=8.2.0
pytz>=2024.1
pyyaml==6.0.1
requests-oauthlib==2.0.0
requests==2.31.0
sqlalchemy==2.0.31
sqlmodel==0.0.19
twitchio==2.9.1
uvicorn==0.29.0
werkzeug==2.3.3
